move code into rest.debug package
